# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /progress/{day}:
    get:
      operationId: getDayProgress
      tags: [progress]
      summary: Get the current user's progress for a day
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Day progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayProgress"
        "404":
          description: Not started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /progress/{day}/lesson:
    post:
      operationId: completeLesson
      tags: [progress]
      summary: Mark the day's lesson complete
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated progress and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressUpdate"
  /progress/{day}/task:
    post:
      operationId: completeTask
      tags: [progress]
      summary: Submit the day's task work for structured AI mentor feedback
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskSubmissionInput"
      responses:
        "200":
          description: Structured AI feedback with updated progress and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskEvalResult"
        "404":
          description: Curriculum day not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /progress/{day}/practice:
    post:
      operationId: submitPractice
      tags: [progress]
      summary: Submit the inline practice exercise for AI grading
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeSubmissionInput"
      responses:
        "200":
          description: Pass/retry result with AI feedback and updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeResult"
        "404":
          description: Curriculum day not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /progress/{day}/quiz:
    post:
      operationId: submitQuiz
      tags: [progress]
      summary: Submit quiz answers for the day
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result with updated progress and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /progress/{day}/reflection:
    post:
      operationId: submitReflection
      tags: [progress]
      summary: Submit a reflection for the day
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionInput"
      responses:
        "200":
          description: Updated progress and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressUpdate"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List the current user's projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Me:
      type: object
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        currentDay:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streakCount:
          type: integer
        streakFreezes:
          type: integer
        weeklyXp:
          type: integer
        isPublicProfile:
          type: boolean
        hideFromLeaderboard:
          type: boolean
        anonymousMode:
          type: boolean
        emailRemindersEnabled:
          type: boolean
        lastActiveDate:
          type: ["string", "null"]
        theme:
          type: string
        hasPaid:
          type: boolean
        isGrandfathered:
          type: boolean
        hasAccess:
          type: boolean
          description: True if the user can access all 28 days (paid or grandfathered).
        accessibleThroughDay:
          type: integer
          description: Highest day number the user can currently open (1 for free preview, 28 if entitled).
        onboardingCompleted:
          type: boolean
        workContext:
          type: ["string", "null"]
        programTitle:
          type: ["string", "null"]
        programFocus:
          type: ["string", "null"]
        dailyMinutes:
          type: integer
          description: Target minutes the learner wants to spend per day (10, 20, or 40). Drives lesson depth and day-completion requirements.
        field:
          type: ["string", "null"]
          description: The domain the learner wants to apply AI to (e.g. marketing, engineering).
        experienceLevel:
          type: ["string", "null"]
          description: Self-reported AI experience (beginner, some, advanced).
        goal:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - currentDay
        - dailyMinutes
        - xp
        - level
        - streakCount
        - streakFreezes
        - weeklyXp
        - isPublicProfile
        - hideFromLeaderboard
        - anonymousMode
        - emailRemindersEnabled
        - theme
        - hasPaid
        - isGrandfathered
        - hasAccess
        - accessibleThroughDay
        - onboardingCompleted
        - createdAt
...
    DayProgress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        day:
          type: integer
        lessonCompleted:
          type: boolean
        practiceCompleted:
          type: boolean
        practiceText:
          type: ["string", "null"]
        taskCompleted:
          type: boolean
        taskSubmission:
          type: ["string", "null"]
        taskScore:
          type: ["integer", "null"]
        taskFeedback:
          oneOf:
            - $ref: "#/components/schemas/TaskFeedback"
            - type: "null"
        quizCompleted:
          type: boolean
        quizScore:
          type: ["integer", "null"]
        reflectionText:
          type: ["string", "null"]
        reflectionCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - day
        - lessonCompleted
        - practiceCompleted
        - taskCompleted
        - quizCompleted
        - reflectionCompleted
        - createdAt
        - updatedAt
    TaskFeedback:
      type: object
      properties:
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        summary:
          type: string
      required:
        - strengths
        - improvements
        - summary
    TaskSubmissionInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 8000
      required:
        - text
    TaskEvalResult:
      type: object
      properties:
        passed:
          type: boolean
        score:
          type: integer
        feedback:
          $ref: "#/components/schemas/TaskFeedback"
        progress:
          $ref: "#/components/schemas/DayProgress"
        user:
          $ref: "#/components/schemas/Me"
        xpAwarded:
          type: integer
        leveledUp:
          type: boolean
        dayCompleted:
          type: boolean
        streakExtended:
          type: boolean
        unlockedAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required:
        - passed
        - score
        - feedback
        - progress
        - user
        - xpAwarded
        - leveledUp
        - dayCompleted
        - streakExtended
        - unlockedAchievements
    PracticeSubmissionInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - text
    PracticeResult:
      type: object
      properties:
        passed:
          type: boolean
        feedback:
          type: string
        progress:
          $ref: "#/components/schemas/DayProgress"
        user:
          $ref: "#/components/schemas/Me"
        xpAwarded:
          type: integer
        leveledUp:
          type: boolean
        dayCompleted:
          type: boolean
        streakExtended:
          type: boolean
        unlockedAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required:
        - passed
        - feedback
        - progress
        - user
        - xpAwarded
        - leveledUp
        - dayCompleted
        - streakExtended
        - unlockedAchievements
...
    Achievement:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - code
        - title
        - description
        - icon
...
    ProgressUpdate:
      type: object
      properties:
        progress:
          $ref: "#/components/schemas/DayProgress"
        user:
          $ref: "#/components/schemas/Me"
        xpAwarded:
          type: integer
        leveledUp:
          type: boolean
        dayCompleted:
          type: boolean
        streakExtended:
          type: boolean
        unlockedAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required:
        - progress
        - user
        - xpAwarded
        - leveledUp
        - dayCompleted
        - streakExtended
        - unlockedAchievements
    QuizSubmission:
      type: object
      properties:
        answers:
          type: array
          items:
            type: integer
      required:
        - answers
    QuizResult:
      type: object
      properties:
        score:
          type: integer
        total:
          type: integer
        passed:
          type: boolean
        progress:
          $ref: "#/components/schemas/DayProgress"
        user:
          $ref: "#/components/schemas/Me"
        xpAwarded:
          type: integer
        leveledUp:
          type: boolean
        dayCompleted:
          type: boolean
        streakExtended:
          type: boolean
        correctIndexes:
          type: array
          items:
            type: integer
        unlockedAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required:
        - score
        - total
        - passed
        - progress
        - user
        - xpAwarded
        - leveledUp
        - dayCompleted
        - streakExtended
        - correctIndexes
        - unlockedAchievements
    ReflectionInput:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - status
        - progress
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
        progress:
          type: integer
      required:
        - title
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
        progress:
          type: integer
...